[ty] Move tests and type-alias-related code out of `types.rs`
