+++ crates/ty_python_semantic/src/types/display.rs
@@ -22,6 +22,7 @@ use crate::place::{DefinedPlace, Place};
 use crate::semantic_index::definition::Definition;
 use crate::semantic_index::scope::{FileScopeId, ScopeKind};
 use crate::semantic_index::semantic_index;
+use crate::types::callable::CallableTypeKind;
 use crate::types::class::{ClassLiteral, ClassType, GenericAlias};
 use crate::types::function::{FunctionType, OverloadLiteral};
 use crate::types::generics::{GenericContext, Specialization};
@@ -32,11 +33,10 @@ use crate::types::tuple::TupleSpec;
 use crate::types::typevar::BoundTypeVarIdentity;
 use crate::types::visitor::TypeVisitor;
 use crate::types::{
-    BindingContext, CallableType, CallableTypeKind, IntersectionType, KnownBoundMethodType,
-    KnownClass, KnownInstanceType, LiteralValueType, LiteralValueTypeKind, MaterializationKind,
-    Protocol, ProtocolInstanceType, SpecialFormType, StringLiteralType, SubclassOfInner,
-    SubclassOfType, Type, TypeAliasType, TypeGuardLike, TypedDictType, UnionType,
-    WrapperDescriptorKind, visitor,
+    BindingContext, CallableType, IntersectionType, KnownBoundMethodType, KnownClass,
+    KnownInstanceType, LiteralValueType, LiteralValueTypeKind, MaterializationKind, Protocol,
+    ProtocolInstanceType, SpecialFormType, StringLiteralType, SubclassOfInner, SubclassOfType,
+    Type, TypeAliasType, TypeGuardLike, TypedDictType, UnionType, WrapperDescriptorKind, visitor,
 };
 
 /// A named item that can be either a class or a type alias.
@@ -646,8 +646,8 @@ fn fmt_file_location<'db>(
 /// Returns the qualified name components for a scope, excluding the item itself.
 ///
 /// This is the shared logic used by both [`QualifiedClassName`](super::class::QualifiedClassName)
-/// and [`QualifiedTypeAliasName`](super::QualifiedTypeAliasName) to compute the path components
-/// (module, enclosing classes, functions) leading to an item.
+/// and [`QualifiedTypeAliasName`](super::type_alias::QualifiedTypeAliasName) to compute the path
+/// components (module, enclosing classes, functions) leading to an item.
 ///
 /// # Returns
 /// A vector of path components in order (e.g., `["module", "OuterClass", "InnerClass"]`)
