+++ crates/ty_python_semantic/src/types/visitor.rs
@@ -15,10 +15,12 @@ use crate::{
         known_instance::walk_known_instance_type,
         method::{walk_bound_method_type, walk_method_wrapper_type},
         newtype::{NewType, walk_newtype_instance_type},
+        set_theoretic::{walk_intersection_type, walk_union},
         subclass_of::walk_subclass_of_type,
+        type_alias::walk_type_alias_type,
+        typed_dict::walk_typed_dict_type,
         typevar::{TypeVarInstance, walk_bound_type_var_type, walk_type_var_type},
-        walk_intersection_type, walk_property_instance_type, walk_type_alias_type,
-        walk_typed_dict_type, walk_typeguard_type, walk_typeis_type, walk_union,
+        walk_property_instance_type, walk_typeguard_type, walk_typeis_type,
     },
 };
 use std::cell::{Cell, RefCell};
